requirements.txt
blinker==1.8.2
Brotli==1.1.0
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
cssselect2==0.7.0
Flask==3.0.3
Flask-WTF==1.2.2
fonttools==4.55.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
more-itertools==10.5.0
numpy==2.1.3
pillow==10.4.0
psutil==6.1.0
pycparser==2.22
pydyf==0.11.0
pyphen==0.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
reportlab==4.2.5
requests==2.32.3
sib-api-v3-sdk==7.6.0
six==1.17.0
tinycss2==1.4.0
tinyhtml5==2.0.0
urllib3==2.2.3
waitress==3.0.0
webencodings==0.5.1
Werkzeug==3.0.4
WTForms==3.1.2
zopfli==0.2.3.post1
